requirements.txt
Pillow>=10.1.0
requests>=2.31.0 